Added HBL2_API macro to engine files.
