+++ Humble2/src/Humble2/Core/Window.h
@@ -11,7 +11,7 @@
 
 namespace HBL2
 {
-	struct WindowSpecification
+	struct HBL2_API WindowSpecification
 	{
 		std::string Title = "Humble App";
 		float RefreshRate = 0.0f;
@@ -21,11 +21,11 @@ namespace HBL2
 		bool VerticalSync = false;
 	};
 
-	class Window
+	class HBL2_API Window
 	{
 	public:
 		~Window() = default;
-		static inline Window* Instance;
+		static Window* Instance;
 
 		void Initialize(const WindowSpecification&& spec);
 
@@ -37,7 +37,7 @@ namespace HBL2
 		void SetTitle(const std::string& title);
 		double GetTime();
 		GLFWwindow* GetHandle();
-		glm::u32vec2 GetExtents() { return { m_Spec.Width, m_Spec.Height }; }
+		const glm::u32vec2& GetExtents() const { return { m_Spec.Width, m_Spec.Height }; }
 
 	protected:
 		void AttachEventCallbacks();
