+++ Humble2/src/Humble2/ImGui/ImGuiRenderer.cpp
@@ -1,0 +1,6 @@
+#include "ImGuiRenderer.h"
+
+namespace HBL2
+{
+	ImGuiRenderer* ImGuiRenderer::Instance = nullptr;
+}
