+++ Humble2/src/Humble2/Project/Project.cpp
@@ -5,6 +5,8 @@
 
 namespace HBL2
 {
+	Project* Project::s_ActiveProject = nullptr;
+
 	Project* Project::Create(const std::string& name)
 	{
 		s_ActiveProject = new Project;
