+++ Humble2/src/Humble2/Renderer/Device.cpp
@@ -1,0 +1,6 @@
+#include "Device.h"
+
+namespace HBL2
+{
+	Device* Device::Instance = nullptr;
+}
