+++ Humble2/src/Humble2/Resources/ResourceManager.cpp
@@ -1,0 +1,6 @@
+#include "ResourceManager.h"
+
+namespace HBL2
+{
+	ResourceManager* ResourceManager::Instance = nullptr;
+}
