+++ Humble2/src/Humble2/Scene/ISystem.h
@@ -1,26 +1,28 @@
 #pragma once
 
+#include "Base.h"
+
 #include <Core\Events.h>
 
 namespace HBL2
 {
 	class Scene;
 
-	enum class SystemType
+	enum class HBL2_API SystemType
 	{
 		Core = 0,
 		Runtime,
 		User,
 	};
 
-	enum class SystemState
+	enum class HBL2_API SystemState
 	{
 		Play = 0,
 		Pause,
 		Idle,
 	};
 
-	class ISystem
+	class HBL2_API ISystem
 	{
 	public:
 		virtual ~ISystem() = default;
