+++ Humble2/src/Humble2/Utilities/YamlUtilities.h
@@ -7,7 +7,7 @@
 namespace YAML
 {
 	template<>
-	struct convert<glm::vec2>
+	struct HBL2_API convert<glm::vec2>
 	{
 		static Node encode(const glm::vec2& rhs)
 		{
@@ -30,7 +30,7 @@ namespace YAML
 	};
 
 	template<>
-	struct convert<glm::vec3>
+	struct HBL2_API convert<glm::vec3>
 	{
 		static Node encode(const glm::vec3& rhs)
 		{
@@ -55,7 +55,7 @@ namespace YAML
 	};
 
 	template<>
-	struct convert<glm::vec4>
+	struct HBL2_API convert<glm::vec4>
 	{
 		static Node encode(const glm::vec4& rhs)
 		{
@@ -84,9 +84,9 @@ namespace YAML
 
 namespace HBL2
 {
-	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec2& v);
+	HBL2_API YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec2& v);
 
-	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec3& v);
+	HBL2_API YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec3& v);
 
-	YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec4& v);
+	HBL2_API YAML::Emitter& operator<<(YAML::Emitter& out, const glm::vec4& v);
 }
