+++ Humble2/src/Humble2API.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#ifdef BUILD_DLL
+	#define HBL2_API __declspec(dllexport)
+#else
+	#define HBL2_API 
+#endif
